# docs/openapi/course-module.openapi.yaml
openapi: 3.1.0
info:
  title: 在线教育辅助教学系统 - 课程与学习基础模块
  version: 0.2.0
  description: 阶段 2 联调契约。所有 Long ID 以字符串传输，时间使用 RFC 3339。
servers:
  - url: http://localhost:8080
    description: 通过 edu-gateway
tags:
  - {name: 认证}
  - {name: 教师课程}
  - {name: 课程审核}
  - {name: 学生课程}
  - {name: 教师内容}
  - {name: 教师资料}
  - {name: 学生学习}
paths:
  /api/v1/auth/login:
    post: {tags: [认证], operationId: login, security: [], requestBody: {$ref: '#/components/requestBodies/Login'}, responses: {'200': {$ref: '#/components/responses/Success'}}}
  /api/v1/teacher/courses:
    get: {tags: [教师课程], operationId: listTeacherCourses, parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}, {$ref: '#/components/parameters/Keyword'}, {$ref: '#/components/parameters/Status'}, {$ref: '#/components/parameters/Term'}, {$ref: '#/components/parameters/CategoryId'}, {$ref: '#/components/parameters/Sort'}], responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}}
    post: {tags: [教师课程], operationId: createCourse, requestBody: {$ref: '#/components/requestBodies/CreateCourse'}, responses: {'201': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [教师课程], operationId: getTeacherCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
    put: {tags: [教师课程], operationId: updateCourse, requestBody: {$ref: '#/components/requestBodies/UpdateCourse'}, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/submit-review:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [教师课程], operationId: submitCourseReview, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/publish:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [教师课程], operationId: publishCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/offline:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [教师课程], operationId: offlineCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/teachers:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [教师课程], operationId: listCourseTeachers, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
    post: {tags: [教师课程], operationId: addCourseTeacher, requestBody: {$ref: '#/components/requestBodies/AddTeacher'}, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/teachers/{teacherId}:
    parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/TeacherId'}]
    delete: {tags: [教师课程], operationId: removeCourseTeacher, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/admin/course-reviews:
    get: {tags: [课程审核], operationId: listCourseReviews, parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}, {$ref: '#/components/parameters/Keyword'}], responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/admin/course-reviews/{courseId}:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [课程审核], operationId: getCourseReview, responses: {'200': {$ref: '#/components/responses/Success'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/admin/course-reviews/{courseId}/approve:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [课程审核], operationId: approveCourse, requestBody: {$ref: '#/components/requestBodies/ApproveCourse'}, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/admin/course-reviews/{courseId}/reject:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [课程审核], operationId: rejectCourse, requestBody: {$ref: '#/components/requestBodies/RejectCourse'}, responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/courses/catalog:
    get: {tags: [学生课程], operationId: listCourseCatalog, parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}, {$ref: '#/components/parameters/Keyword'}, {$ref: '#/components/parameters/Term'}, {$ref: '#/components/parameters/CategoryId'}, {$ref: '#/components/parameters/Sort'}], responses: {'200': {$ref: '#/components/responses/Success'}}}
  /api/v1/student/courses:
    get: {tags: [学生课程], operationId: listStudentCourses, parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}, {$ref: '#/components/parameters/Keyword'}, {$ref: '#/components/parameters/Status'}, {$ref: '#/components/parameters/Term'}, {$ref: '#/components/parameters/CategoryId'}, {$ref: '#/components/parameters/Sort'}], responses: {'200': {$ref: '#/components/responses/Success'}}}
  /api/v1/student/courses/{courseId}:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [学生课程], operationId: getStudentCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/courses/{courseId}/enroll:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [学生课程], operationId: enrollCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/courses/{courseId}/withdraw:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    post: {tags: [学生课程], operationId: withdrawCourse, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/chapters:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [教师内容], operationId: listChapters, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
    post: {tags: [教师内容], operationId: createChapter, requestBody: {$ref: '#/components/requestBodies/CreateChapter'}, responses: {'201': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/chapters/{chapterId}:
    parameters: [{$ref: '#/components/parameters/ChapterId'}]
    put: {tags: [教师内容], operationId: updateChapter, requestBody: {$ref: '#/components/requestBodies/UpdateChapter'}, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
    delete: {tags: [教师内容], operationId: deleteChapter, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/chapters/{chapterId}/publish:
    parameters: [{$ref: '#/components/parameters/ChapterId'}]
    post: {tags: [教师内容], operationId: publishChapter, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/chapters/{chapterId}/offline:
    parameters: [{$ref: '#/components/parameters/ChapterId'}]
    post: {tags: [教师内容], operationId: offlineChapter, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/chapters/{chapterId}/lessons:
    parameters: [{$ref: '#/components/parameters/ChapterId'}]
    get: {tags: [教师内容], operationId: listLessons, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
    post: {tags: [教师内容], operationId: createLesson, requestBody: {$ref: '#/components/requestBodies/CreateLesson'}, responses: {'201': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/lessons/{lessonId}:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    get: {tags: [教师内容], operationId: getTeacherLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
    put: {tags: [教师内容], operationId: updateLesson, requestBody: {$ref: '#/components/requestBodies/UpdateLesson'}, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
    delete: {tags: [教师内容], operationId: deleteLesson, responses: {'200': {$ref: '#/components/responses/Success'}}}
  /api/v1/teacher/lessons/{lessonId}/publish:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    post: {tags: [教师内容], operationId: publishLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/lessons/{lessonId}/offline:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    post: {tags: [教师内容], operationId: offlineLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/courses/{courseId}/materials:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [教师资料], operationId: listMaterials, parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}, {$ref: '#/components/parameters/Keyword'}, {$ref: '#/components/parameters/Status'}], responses: {'200': {$ref: '#/components/responses/Success'}}}
    post: {tags: [教师资料], operationId: createMaterial, requestBody: {$ref: '#/components/requestBodies/CreateMaterial'}, responses: {'201': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/teacher/materials/{materialId}:
    parameters: [{$ref: '#/components/parameters/MaterialId'}]
    put: {tags: [教师资料], operationId: updateMaterial, requestBody: {$ref: '#/components/requestBodies/UpdateMaterial'}, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}}}
    delete: {tags: [教师资料], operationId: deleteMaterial, responses: {'200': {$ref: '#/components/responses/Success'}}}
  /api/v1/student/courses/{courseId}/outline:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [学生学习], operationId: getCourseOutline, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/lessons/{lessonId}:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    get: {tags: [学生学习], operationId: getStudentLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/lessons/{lessonId}/start:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    post: {tags: [学生学习], operationId: startLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/lessons/{lessonId}/complete:
    parameters: [{$ref: '#/components/parameters/LessonId'}]
    post: {tags: [学生学习], operationId: completeLesson, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/courses/{courseId}/progress:
    parameters: [{$ref: '#/components/parameters/CourseId'}]
    get: {tags: [学生学习], operationId: getCourseProgress, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/student/materials/{materialId}:
    parameters: [{$ref: '#/components/parameters/MaterialId'}]
    get: {tags: [学生学习], operationId: getMaterialAccess, responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    CourseId: {name: courseId, in: path, required: true, schema: {type: string, pattern: '^[0-9]+$'}}
    TeacherId: {name: teacherId, in: path, required: true, schema: {type: string, pattern: '^[0-9]+$'}}
    ChapterId: {name: chapterId, in: path, required: true, schema: {type: string, pattern: '^[0-9]+$'}}
    LessonId: {name: lessonId, in: path, required: true, schema: {type: string, pattern: '^[0-9]+$'}}
    MaterialId: {name: materialId, in: path, required: true, schema: {type: string, pattern: '^[0-9]+$'}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    Size: {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Keyword: {name: keyword, in: query, schema: {type: string}}
    Status: {name: status, in: query, schema: {type: string}}
    Term: {name: term, in: query, schema: {type: string}}
    CategoryId: {name: categoryId, in: query, schema: {type: string}}
    Sort: {name: sort, in: query, schema: {type: string, default: 'createdAt,desc'}}
  requestBodies:
    Login: {$ref: '#/components/requestBodies/LoginBody'}
    LoginBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}, examples: {teacher: {value: {username: teacher, password: 'Teacher@123'}}}}}}
    CreateCourse: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CourseWriteRequest'}}}}
    UpdateCourse: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CourseWriteRequest'}}}}
    AddTeacher: {required: true, content: {application/json: {schema: {type: object, required: [teacherId], properties: {teacherId: {type: string, example: '1004'}}}}}}
    ApproveCourse: {required: false, content: {application/json: {schema: {type: object, properties: {remark: {type: string, maxLength: 500}}}}}}
    RejectCourse: {required: true, content: {application/json: {schema: {type: object, required: [reason], properties: {reason: {type: string, minLength: 1, maxLength: 500}}}}}}
    CreateChapter: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChapterWriteRequest'}}}}
    UpdateChapter: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChapterWriteRequest'}}}}
    CreateLesson: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LessonWriteRequest'}}}}
    UpdateLesson: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LessonWriteRequest'}}}}
    CreateMaterial: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MaterialWriteRequest'}}}}
    UpdateMaterial: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MaterialWriteRequest'}}}}
  responses:
    Success: {description: 统一成功响应, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
    Error: {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string, format: password}}
    CourseWriteRequest:
      type: object
      required: [name]
      properties:
        courseCode: {type: string, maxLength: 64, description: 仅创建时必填}
        name: {type: string, maxLength: 160}
        summary: {type: string, maxLength: 5000}
        term: {type: string, maxLength: 32}
        department: {type: string, maxLength: 128}
        credit: {type: number}
        enrollmentOpenAt: {type: string, format: date-time}
        enrollmentCloseAt: {type: string, format: date-time}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        version: {type: integer, minimum: 0, description: 仅更新时必填}
    ChapterWriteRequest:
      type: object
      required: [title, sortOrder]
      properties: {title: {type: string, maxLength: 120}, description: {type: string, maxLength: 2000}, sortOrder: {type: integer, minimum: 0}, version: {type: integer, minimum: 0}}
    LessonWriteRequest:
      type: object
      required: [title, contentType, sortOrder, unlockType]
      properties:
        courseId: {type: string, description: 可选上下文断言，不改变服务端归属}
        title: {type: string, maxLength: 160}
        contentType: {type: string, enum: [RICH_TEXT, VIDEO, DOCUMENT, MIXED]}
        content: {type: string, description: Markdown}
        videoUrl: {type: string}
        estimatedMinutes: {type: integer, minimum: 1}
        sortOrder: {type: integer, minimum: 0}
        unlockType: {type: string, enum: [IMMEDIATE, SCHEDULED]}
        unlockAt: {type: string, format: date-time}
        version: {type: integer, minimum: 0}
    MaterialWriteRequest:
      type: object
      required: [name, materialType, visibility, sortOrder]
      properties:
        chapterId: {type: string}
        lessonId: {type: string}
        name: {type: string, maxLength: 160}
        materialType: {type: string, enum: [COURSEWARE, DOCUMENT, VIDEO, CODE, LINK, OTHER]}
        fileKey: {type: string, description: 仅教师端元数据}
        fileUrl: {type: string}
        fileSize: {type: integer, minimum: 0}
        mimeType: {type: string}
        visibility: {type: string, enum: [COURSE, CHAPTER, LESSON]}
        status: {type: string, enum: [DRAFT, PUBLISHED, OFFLINE]}
        sortOrder: {type: integer, minimum: 0}
        version: {type: integer, minimum: 0}
    ApiResponse:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code: {type: string, example: SUCCESS}
        message: {type: string, example: OK}
        data: {}
        errors: {type: array, items: {type: object}}
        traceId: {type: string}
        timestamp: {type: string, format: date-time}
security:
  - bearerAuth: []